Split Diagnostics.swift into multiple files
No functional changes here; simply splits the types defined in
Diagnostics.swift into multiple files with one type per file. I
anticipate some clean up related to diagnostics and argument parsing to
occur in a further PR and this change will make future changes easier to
grok.
