+++ Sources/MMIOMacros/SwiftSyntaxExtensions/Diagnostics/ExpansionError.swift
@@ -1,0 +1,13 @@
+//===----------------------------------------------------------*- swift -*-===//
+//
+// This source file is part of the Swift MMIO open source project
+//
+// Copyright (c) 2023 Apple Inc. and the Swift project authors
+// Licensed under Apache License v2.0 with Runtime Library Exception
+//
+// See https://swift.org/LICENSE.txt for license information
+//
+//===----------------------------------------------------------------------===//
+
+/// A marker error used as an early exit for a failed macro expansion.
+struct ExpansionError: Error {}
